pyproject.toml
[project]
name = "priorauth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "annotated-types==0.7.0",
    "certifi==2026.1.4",
    "charset-normalizer==3.4.4",
    "chromadb>=1.5.5",
    "google-genai>=1.69.0",
    "google-generativeai>=0.8.6",
    "idna==3.11",
    "isodate==0.7.2",
    "langchain>=1.2.14",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-huggingface>=1.2.1",
    "langgraph>=1.1.4",
    "mistralai>=2.2.0",
    "openai>=2.30.0",
    "paddleocr>=3.4.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pymupdf>=1.27.2.2",
    "pypdf>=4.0.0",
    "python-dotenv==1.2.1",
    "requests==2.32.5",
    "sentence-transformers>=5.3.0",
    "torch>=2.11.0",
    "transformers>=5.5.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
]
